- refactor the post processing graph code into new dir - fix bug in the r calculation
